+++ Source/RebalancePatches/Mods/AlteredCarbon/CastingRelayRangeExtension.cs
@@ -1,0 +1,9 @@
+using Verse;
+
+namespace RebalancePatches
+{
+    public class CastingRelayRangeExtension : DefModExtension
+    {
+        public int tilesPerRelay = 5;
+    }
+}